Proposal: Blood Hunt simplification
Proposal after long debates on wether a Blood Hunt is during an action.

The current wording read like reactions and modifiers cannot be used
because there is no action.

This is not the case, though. The Blood Hunt cannot be blocked,
but it takes place during an action.

Modifiers and reactions cannot be played,
but not because the blood hunt "window" is somehow "outside the action".
You cannot use Heidelberg during the Blood Hunt (ongoing action).

It is simpler to just rule that you cannot use modifiers or reactions.
It makes Ventrue Headquarters interaction clearer.

**Note this is a functional change for Vele's Hunt.**